Basic Input Checking

+++ hillClimb.h
@@ -1,5 +1,5 @@
 #ifndef HILLCLIMB_H
 #define HILLCLIMB_H
 #include "board.h"
-ChessBoard hillClimb (int row, int col, int white, int black, int tmax);
+ChessBoard hillClimb (int row, int col, int white, int black, double tmax);
 #endif
